3 sprint

+++ 2 Инструменты разработчика/2.3.3 Аннотация типов задача.py
@@ -6,7 +6,7 @@ def add_numbers(x_1: int, y_1: int) -> int:
     return x_1 + y_1
 
 
-def calculate_square_root(number: int) -> float:
+def calculate_square_root(number: int) -> int:
     return sqrt(number)
